Move interface Resource under lib/types
