+++ src/resource/page/ResourcePage.tsx
@@ -1,7 +1,7 @@
 import React, { ReactNode, useEffect, useState } from 'react';
 import { Notification } from 'hds-react';
-import api, { Resource } from '../../common/utils/api/api';
-import { DatePeriod } from '../../common/lib/types';
+import api from '../../common/utils/api/api';
+import { DatePeriod, Resource } from '../../common/lib/types';
 import Collapse from '../../components/collapse/Collapse';
 import { ExternalLink } from '../../components/link/Link';
 import ResourceOpeningHours from '../resource-opening-hours/ResourceOpeningHours';
